biddingproject/requirements.txt
asgiref==3.7.2
Django==4.2.3
django-allauth==0.52.0
django-cors-headers==4.2.0
djangorestframework==3.14.0
Pillow==10.0.0
soupsieve==2.4.1
whitenoise==6.5.0
dj-rest-auth
drf_yasg